+++ example/lib/main.dart
@@ -1,64 +1,606 @@
-import 'package:flutter/material.dart';
-
 import 'package:awesome_notifications/awesome_notifications.dart';
-import 'package:awesome_notifications_fcm_example/routes.dart';
+import 'package:awesome_notifications_fcm/awesome_notifications_fcm.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter/material.dart';
+import 'package:http/http.dart' as http;
+import 'package:fluttertoast/fluttertoast.dart';
 
-import 'notifications/notification_controller.dart';
-
-main() async {
-  WidgetsFlutterBinding.ensureInitialized();
+Future<void> main() async {
   await NotificationController.initializeLocalNotifications(debug: true);
   await NotificationController.initializeRemoteNotifications(debug: true);
   await NotificationController.getInitialNotificationAction();
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
+
+
+
+///  *********************************************
+///     NOTIFICATION CONTROLLER
+///  *********************************************
+
+class NotificationController extends ChangeNotifier {
+  /// *********************************************
+  ///   SINGLETON PATTERN
+  /// *********************************************
+
+  static final NotificationController _instance =
+      NotificationController._internal();
+
+  factory NotificationController() {
+    return _instance;
+  }
+
+  NotificationController._internal();
+
+  /// *********************************************
+  ///  OBSERVER PATTERN
+  /// *********************************************
+
+  String _firebaseToken = '';
+  String get firebaseToken => _firebaseToken;
+
+  String _nativeToken = '';
+  String get nativeToken => _nativeToken;
+
+  ReceivedAction? initialAction;
+
+  /// *********************************************
+  ///   INITIALIZATION METHODS
+  /// *********************************************
+
+  static Future<void> initializeLocalNotifications(
+      {required bool debug}) async {
+    await AwesomeNotifications().initialize(
+        null, //'resource://drawable/res_app_icon',//
+        [
+          NotificationChannel(
+              channelKey: 'alerts',
+              channelName: 'Alerts',
+              channelDescription: 'Notification tests as alerts',
+              playSound: true,
+              importance: NotificationImportance.High,
+              defaultPrivacy: NotificationPrivacy.Private,
+              defaultColor: Colors.deepPurple,
+              ledColor: Colors.deepPurple)
+        ],
+        debug: debug);
+
+    // Get initial notification action is optional
+    _instance.initialAction = await AwesomeNotifications()
+        .getInitialNotificationAction(removeFromActionEvents: false);
+  }
+
+  static Future<void> initializeRemoteNotifications(
+      {required bool debug}) async {
+    await Firebase.initializeApp();
+    await AwesomeNotificationsFcm().initialize(
+        onFcmSilentDataHandle: NotificationController.mySilentDataHandle,
+        onFcmTokenHandle: NotificationController.myFcmTokenHandle,
+        onNativeTokenHandle: NotificationController.myNativeTokenHandle,
+        licenseKeys:
+        // On this example app, the app ID / Bundle Id are different
+        // for each platform, so i used the main Bundle ID + 1 variation
+        [
+            // me.carda.awesomeNotificationsFcmExample
+            'B3J3yxQbzzyz0KmkQR6rDlWB5N68sTWTEMV7k9HcPBroUh4RZ/Og2Fv6Wc/lE'
+            '2YaKuVY4FUERlDaSN4WJ0lMiiVoYIRtrwJBX6/fpPCbGNkSGuhrx0Rekk'
+            '+yUTQU3C3WCVf2D534rNF3OnYKUjshNgQN8do0KAihTK7n83eUD60=',
+
+            // me.carda.awesome_notifications_fcm_example
+            'UzRlt+SJ7XyVgmD1WV+7dDMaRitmKCKOivKaVsNkfAQfQfechRveuKblFnCp4'
+            'zifTPgRUGdFmJDiw1R/rfEtTIlZCBgK3Wa8MzUV4dypZZc5wQIIVsiqi0Zhaq'
+            'YtTevjLl3/wKvK8fWaEmUxdOJfFihY8FnlrSA48FW94XWIcFY=',
+        ],
+        debug: debug);
+  }
+
+  static Future<void> startListeningNotificationEvents() async {
+    AwesomeNotifications()
+        .setListeners(onActionReceivedMethod: onActionReceivedMethod);
+  }
+
+  ///  *********************************************
+  ///     LOCAL NOTIFICATION EVENTS
+  ///  *********************************************
+
+  static Future<void> getInitialNotificationAction() async {
+    ReceivedAction? receivedAction = await AwesomeNotifications()
+        .getInitialNotificationAction(removeFromActionEvents: true);
+    if (receivedAction == null) return;
+
+    // Fluttertoast.showToast(
+    //     msg: 'Notification action launched app: $receivedAction',
+    //   backgroundColor: Colors.deepPurple
+    // );
+    print('Notification action launched app: $receivedAction');
+  }
+
+  @pragma('vm:entry-point')
+  static Future<void> onActionReceivedMethod(
+      ReceivedAction receivedAction) async {
+    MyApp.navigatorKey.currentState?.pushNamedAndRemoveUntil(
+        '/notification-page',
+        (route) =>
+            (route.settings.name != '/notification-page') || route.isFirst,
+        arguments: receivedAction);
+  }
+
+  ///  *********************************************
+  ///     REMOTE NOTIFICATION EVENTS
+  ///  *********************************************
+
+  /// Use this method to execute on background when a silent data arrives
+  /// (even while terminated)
+  @pragma("vm:entry-point")
+  static Future<void> mySilentDataHandle(FcmSilentData silentData) async {
+    Fluttertoast.showToast(
+        msg: 'Silent data received',
+        backgroundColor: Colors.blueAccent,
+        textColor: Colors.white,
+        fontSize: 16);
+
+    print('"SilentData": ${silentData.toString()}');
+
+    if (silentData.createdLifeCycle != NotificationLifeCycle.Foreground) {
+      print("bg");
+    } else {
+      print("FOREGROUND");
+    }
+
+    print('mySilentDataHandle received a FcmSilentData execution');
+    await executeLongTaskInBackground();
+  }
+
+  /// Use this method to detect when a new fcm token is received
+  @pragma("vm:entry-point")
+  static Future<void> myFcmTokenHandle(String token) async {
+    Fluttertoast.showToast(
+        msg: 'Fcm token received',
+        backgroundColor: Colors.blueAccent,
+        textColor: Colors.white,
+        fontSize: 16);
+    debugPrint('Firebase Token:"$token"');
+
+    _instance._firebaseToken = token;
+    _instance.notifyListeners();
+  }
+
+  /// Use this method to detect when a new native token is received
+  @pragma("vm:entry-point")
+  static Future<void> myNativeTokenHandle(String token) async {
+    Fluttertoast.showToast(
+        msg: 'Native token received',
+        backgroundColor: Colors.blueAccent,
+        textColor: Colors.white,
+        fontSize: 16);
+    debugPrint('Native Token:"$token"');
+
+    _instance._nativeToken = token;
+    _instance.notifyListeners();
+  }
+
+  ///  *********************************************
+  ///     BACKGROUND TASKS TEST
+  ///  *********************************************
+
+  static Future<void> executeLongTaskInBackground() async {
+    print("starting long task");
+    await Future.delayed(const Duration(seconds: 4));
+    final url = Uri.parse("http://google.com");
+    final re = await http.get(url);
+    print(re.body);
+    print("long task done");
+  }
+
+  ///  *********************************************
+  ///     REQUEST NOTIFICATION PERMISSIONS
+  ///  *********************************************
+
+  static Future<bool> displayNotificationRationale() async {
+    bool userAuthorized = false;
+    BuildContext context = MyApp.navigatorKey.currentContext!;
+    await showDialog(
+        context: context,
+        builder: (BuildContext ctx) {
+          return AlertDialog(
+            title: Text('Get Notified!',
+                style: Theme.of(context).textTheme.titleLarge),
+            content: Column(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                Row(
+                  children: [
+                    Expanded(
+                      child: Image.asset(
+                        'assets/animated-bell.gif',
+                        height: MediaQuery.of(context).size.height * 0.3,
+                        fit: BoxFit.fitWidth,
+                      ),
+                    ),
+                  ],
+                ),
+                const SizedBox(height: 20),
+                const Text(
+                    'Allow Awesome Notifications to send you beautiful notifications!'),
+              ],
+            ),
+            actions: [
+              TextButton(
+                  onPressed: () {
+                    Navigator.of(ctx).pop();
+                  },
+                  child: Text(
+                    'Deny',
+                    style: Theme.of(context)
+                        .textTheme
+                        .titleLarge
+                        ?.copyWith(color: Colors.red),
+                  )),
+              TextButton(
+                  onPressed: () async {
+                    userAuthorized = true;
+                    Navigator.of(ctx).pop();
+                  },
+                  child: Text(
+                    'Allow',
+                    style: Theme.of(context)
+                        .textTheme
+                        .titleLarge
+                        ?.copyWith(color: Colors.deepPurple),
+                  )),
+            ],
+          );
+        });
+    return userAuthorized &&
+        await AwesomeNotifications().requestPermissionToSendNotifications();
+  }
+
+  ///  *********************************************
+  ///     LOCAL NOTIFICATION CREATION METHODS
+  ///  *********************************************
+
+  static Future<void> createNewNotification() async {
+    bool isAllowed = await AwesomeNotifications().isNotificationAllowed();
+
+    if (!isAllowed) {
+      isAllowed = await displayNotificationRationale();
+    }
+
+    if (!isAllowed) return;
+
+    await AwesomeNotifications().createNotification(
+        content: NotificationContent(
+            id: -1, // -1 is replaced by a random number
+            channelKey: 'alerts',
+            title: 'Huston! The eagle has landed!',
+            body:
+                "A small step for a man, but a giant leap to Flutter's community!",
+            bigPicture: 'https://storage.googleapis.com/cms-storage-bucket/d406c736e7c4c57f5f61.png',
+            largeIcon: 'https://storage.googleapis.com/cms-storage-bucket/0dbfcc7a59cd1cf16282.png',
+            notificationLayout: NotificationLayout.BigPicture,
+            payload: {'notificationId': '1234567890'}),
+        actionButtons: [
+          NotificationActionButton(key: 'REDIRECT', label: 'Redirect'),
+          NotificationActionButton(
+              key: 'REPLY',
+              label: 'Reply Message',
+              requireInputText: true,
+              actionType: ActionType.SilentAction
+          ),
+          NotificationActionButton(
+              key: 'DISMISS',
+              label: 'Dismiss',
+              actionType: ActionType.DismissAction,
+              isDangerousOption: true)
+        ]);
+  }
+
+  static Future<void> resetBadge() async {
+    await AwesomeNotifications().resetGlobalBadge();
+  }
+
+  ///  *********************************************
+  ///     REMOTE TOKEN REQUESTS
+  ///  *********************************************
+
+  static Future<String> requestFirebaseToken() async {
+    if (await AwesomeNotificationsFcm().isFirebaseAvailable) {
+      try {
+        return await AwesomeNotificationsFcm().requestFirebaseAppToken();
+      } catch (exception) {
+        debugPrint('$exception');
+      }
+    } else {
+      debugPrint('Firebase is not available on this project');
+    }
+    return '';
+  }
+}
+
+
+
+
+///  *********************************************
+///     MAIN WIDGET
+///  *********************************************
+
 class MyApp extends StatefulWidget {
+  const MyApp({super.key});
+
+  // The navigator key is necessary to navigate using static methods
   static final GlobalKey<NavigatorState> navigatorKey =
       GlobalKey<NavigatorState>();
 
-  static const String name = 'Awesome Notifications Fcm - Example App';
-  static final Color mainColor = Color(0xFF9D50DD);
+  static Color mainColor = const Color(0xFF9D50DD);
 
   @override
-  _MyAppState createState() => _MyAppState();
+  State<MyApp> createState() => _AppState();
 }
 
-class _MyAppState extends State<MyApp> {
+class _AppState extends State<MyApp> {
+  // This widget is the root of your application.
+
+  static const String routeHome = '/', routeNotification = '/notification-page';
 
   @override
   void initState() {
+    NotificationController.startListeningNotificationEvents();
+    NotificationController.requestFirebaseToken();
     super.initState();
-
-    // Only after at least the action method is set, the notification events are delivered
-    NotificationController.initializeNotificationListeners();
   }
 
-  @override
-  void dispose() {
-    AwesomeNotifications().dispose();
-    super.dispose();
+  List<Route<dynamic>> onGenerateInitialRoutes(String initialRouteName) {
+    List<Route<dynamic>> pageStack = [];
+    pageStack.add(MaterialPageRoute(
+        builder: (_) =>
+            const MyHomePage(title: 'Awesome Notifications FCM Example App')));
+    if (initialRouteName == routeNotification &&
+        NotificationController().initialAction != null) {
+      pageStack.add(MaterialPageRoute(
+          builder: (_) => NotificationPage(
+              receivedAction: NotificationController().initialAction!)));
+    }
+    return pageStack;
+  }
+
+  Route<dynamic>? onGenerateRoute(RouteSettings settings) {
+    switch (settings.name) {
+      case routeHome:
+        return MaterialPageRoute(
+            builder: (_) => const MyHomePage(
+                title: 'Awesome Notifications FCM Example App'));
+
+      case routeNotification:
+        ReceivedAction receivedAction = settings.arguments as ReceivedAction;
+        return MaterialPageRoute(
+            builder: (_) => NotificationPage(receivedAction: receivedAction));
+    }
+    return null;
   }
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      title: 'Awesome Notifications - Simple Example',
+      navigatorKey: MyApp.navigatorKey,
+      onGenerateInitialRoutes: onGenerateInitialRoutes,
+      onGenerateRoute: onGenerateRoute,
+      theme: ThemeData(
+        primarySwatch: Colors.deepPurple,
+      ),
+    );
+  }
+}
 
-        // The navigator key is necessary to allow to navigate through static methods
-        navigatorKey: MyApp.navigatorKey,
-        title: 'Awesome Notifications FCM',
-        color: MyApp.mainColor,
-        theme: ThemeData(
-            primaryColor: MyApp.mainColor,
-            appBarTheme: AppBarTheme(
-              brightness: Brightness.light,
-              backgroundColor: Colors.white,
-              iconTheme: IconThemeData(color: MyApp.mainColor),
-              textTheme: TextTheme(
-                  headline6: TextStyle(
-                      color: MyApp.mainColor, fontWeight: FontWeight.bold)),
-            )),
-        initialRoute: PAGE_HOME,
-        routes: materialRoutes);
+
+
+
+
+///  *********************************************
+///     HOME PAGE
+///  *********************************************
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key, required this.title});
+  final String title;
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  @override
+  void initState() {
+    NotificationController().addListener(() => setState(() {}));
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.center,
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Text(
+              'Firebase Token:',
+            ),
+            Padding(
+              padding: const EdgeInsets.all(20.0),
+              child: Text(
+                '${NotificationController().firebaseToken}',
+              ),
+            ),
+            SizedBox(height: 20),
+            Text(
+              'Native Token:',
+            ),
+            Padding(
+              padding: const EdgeInsets.all(20.0),
+              child: Text(
+                '${NotificationController().nativeToken}',
+              ),
+            ),
+            SizedBox(height: 20),
+            Padding(
+              padding: const EdgeInsets.all(20.0),
+              child: const Text(
+                'Push the button to create a new local notification or reset the badge counter',
+              ),
+            ),
+          ],
+        ),
+      ),
+      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
+      floatingActionButton: Padding(
+        padding: const EdgeInsets.all(8.0),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            FloatingActionButton(
+              heroTag: '2',
+              onPressed: () => NotificationController.createNewNotification(),
+              tooltip: 'Create New notification',
+              child: const Icon(Icons.outgoing_mail),
+            ),
+            SizedBox(width: 20),
+            FloatingActionButton(
+              heroTag: '1',
+              onPressed: () => NotificationController.resetBadge(),
+              tooltip: 'Reset Badge',
+              child: const Icon(Icons.exposure_zero),
+            ),
+          ],
+        ),
+      ), // This trailing comma makes auto-formatting nicer for build methods.
+    );
+  }
+}
+
+
+
+
+///  *********************************************
+///     NOTIFICATION PAGE
+///  *********************************************
+
+class NotificationPage extends StatelessWidget {
+  const NotificationPage({Key? key, required this.receivedAction})
+      : super(key: key);
+
+  final ReceivedAction receivedAction;
+
+  @override
+  Widget build(BuildContext context) {
+    bool hasLargeIcon = receivedAction.largeIconImage != null;
+    bool hasBigPicture = receivedAction.bigPictureImage != null;
+    double bigPictureSize = MediaQuery.of(context).size.height * .4;
+    double largeIconSize =
+        MediaQuery.of(context).size.height * (hasBigPicture ? .12 : .2);
+
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(receivedAction.title ?? receivedAction.body ?? ''),
+      ),
+      body: SingleChildScrollView(
+        padding: EdgeInsets.zero,
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.start,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            SizedBox(
+                height:
+                    hasBigPicture ? bigPictureSize + 40 : largeIconSize + 60,
+                child: hasBigPicture
+                    ? Stack(
+                        children: [
+                          if (hasBigPicture)
+                            FadeInImage(
+                              placeholder: const NetworkImage(
+                                  'https://cdn.syncfusion.com/content/images/common/placeholder.gif'),
+                              //AssetImage('assets/images/placeholder.gif'),
+                              height: bigPictureSize,
+                              width: MediaQuery.of(context).size.width,
+                              image: receivedAction.bigPictureImage!,
+                              fit: BoxFit.cover,
+                            ),
+                          if (hasLargeIcon)
+                            Positioned(
+                              bottom: 15,
+                              left: 20,
+                              child: ClipRRect(
+                                borderRadius: BorderRadius.all(
+                                    Radius.circular(largeIconSize)),
+                                child: FadeInImage(
+                                  placeholder: const NetworkImage(
+                                      'https://cdn.syncfusion.com/content/images/common/placeholder.gif'),
+                                  //AssetImage('assets/images/placeholder.gif'),
+                                  height: largeIconSize,
+                                  width: largeIconSize,
+                                  image: receivedAction.largeIconImage!,
+                                  fit: BoxFit.cover,
+                                ),
+                              ),
+                            )
+                        ],
+                      )
+                    : Center(
+                        child: ClipRRect(
+                          borderRadius:
+                              BorderRadius.all(Radius.circular(largeIconSize)),
+                          child: FadeInImage(
+                            placeholder: const NetworkImage(
+                                'https://cdn.syncfusion.com/content/images/common/placeholder.gif'),
+                            //AssetImage('assets/images/placeholder.gif'),
+                            height: largeIconSize,
+                            width: largeIconSize,
+                            image: receivedAction.largeIconImage!,
+                            fit: BoxFit.cover,
+                          ),
+                        ),
+                      )),
+            Padding(
+              padding: const EdgeInsets.only(bottom: 20.0, left: 20, right: 20),
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.start,
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  RichText(
+                      text: TextSpan(children: [
+                    if (receivedAction.title?.isNotEmpty ?? false)
+                      TextSpan(
+                        text: receivedAction.title!,
+                        style: Theme.of(context).textTheme.titleLarge,
+                      ),
+                    if ((receivedAction.title?.isNotEmpty ?? false) &&
+                        (receivedAction.body?.isNotEmpty ?? false))
+                      TextSpan(
+                        text: '\n\n',
+                        style: Theme.of(context).textTheme.bodyMedium,
+                      ),
+                    if (receivedAction.body?.isNotEmpty ?? false)
+                      TextSpan(
+                        text: receivedAction.body!,
+                        style: Theme.of(context).textTheme.bodyMedium,
+                      ),
+                  ]))
+                ],
+              ),
+            ),
+            Container(
+              color: Colors.black12,
+              padding: const EdgeInsets.all(20),
+              width: MediaQuery.of(context).size.width,
+              child: Text(receivedAction.toString()),
+            ),
+          ],
+        ),
+      ),
+    );
   }
 }